@@ -5,10 +5,10 @@ namespace Shared.Models;
 public class Attack
 {
     [JsonPropertyName("name")]
-    public string Name { get; set; }
+    public string Name { get; set; } = string.Empty;
 
     [JsonPropertyName("type")]
-    public string Type { get; set; }
+    public string Type { get; set; } = string.Empty;
 
     [JsonPropertyName("damage")]
     public int Damage { get; set; }
