@@ -5,10 +5,10 @@ namespace Shared.Models;
 public class PokemonDimension
 {
     [JsonPropertyName("minimum")]
-    public string Minimum { get; set; }
+    public string Minimum { get; set; } = string.Empty;
 
     [JsonPropertyName("maximum")]
-    public string Maximum { get; set; }
+    public string Maximum { get; set; } = string.Empty;
 
     public override string ToString()
     {
