@@ -1,4 +1,3 @@
-using System;
 using GraphQL.Query.Builder.UnitTests.Models;
 using Xunit;
 
@@ -9,15 +8,15 @@ public class CamelCasePropertyNameFormatterTests
     [Fact]
     public void Format_ShouldTransfomeNameIntoCamelCase()
     {
-        {
-            string name = CamelCasePropertyNameFormatter.Format.Invoke(typeof(Car).GetProperty(nameof(Car.Name)));
-            Assert.Equal("name", name);
-        }
+        string name = CamelCasePropertyNameFormatter.Format.Invoke(typeof(Car).GetProperty(nameof(Car.Name))!);
+        Assert.Equal("name", name);
     }
 
     [Fact]
     public void Format_ShouldThrowIfPropertyIsNull()
     {
+#nullable disable
         Assert.Throws<ArgumentNullException>(() => CamelCasePropertyNameFormatter.Format.Invoke(null));
+#nullable restore
     }
 }
